refactor(asset) update asset DTOs to use BigDecimal for price and Integer for quantity

+++ src/main/java/com/ivocorrea/investmanager/dto/asset/PutAssetDTO.java
@@ -1,4 +1,6 @@
 package com.ivocorrea.investmanager.dto.asset;
 
-public record PutAssetDTO(int quantity, double currentPrice) {
+import java.math.BigDecimal;
+
+public record PutAssetDTO(Integer quantity, BigDecimal currentPrice) {
 }
